pyproject.toml
[project]
name = "resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "pillow>=10.0.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "reportlab>=4.1.0",
    "streamlit>=1.32.0",
]
